requirements.txt
flask
openai
python-dotenv
psycopg2-binary
requests
beautifulsoup4 
gunicorn